requirements.txt
icalendar
openpyxl
python-dateutil
python-dotenv
rich